backend/requirements.txt
Flask==3.0.3
pandas==2.1.4
flask-cors==4.0.0
matplotlib==3.9.2
arabic_reshaper==3.0.0
python-bidi==0.4.2